Extract the cop classes from if_then_else.rb into separate files
