+++ lib/rubocop.rb
@@ -63,6 +63,9 @@ require 'rubocop/cop/style/favor_sprintf'
 require 'rubocop/cop/style/favor_unless_over_negated_if'
 require 'rubocop/cop/style/hash_syntax'
 require 'rubocop/cop/style/if_then_else'
+require 'rubocop/cop/style/if_with_semicolon'
+require 'rubocop/cop/style/multiline_if_then'
+require 'rubocop/cop/style/one_line_conditional'
 require 'rubocop/cop/style/lambda'
 require 'rubocop/cop/style/leading_comment_space'
 require 'rubocop/cop/style/line_continuation'
