+++ lib/rubocop/cop/style/if_with_semicolon.rb
@@ -1,0 +1,20 @@
+# encoding: utf-8
+
+module Rubocop
+  module Cop
+    module Style
+      # Checks for uses of semicolon in if statements.
+      class IfWithSemicolon < Cop
+        include IfThenElse
+
+        def offending_line(node)
+          node.loc.begin.line if node.loc.begin && node.loc.begin.is?(';')
+        end
+
+        def error_message
+          'Never use if x; Use the ternary operator instead.'
+        end
+      end
+    end
+  end
+end
